Add Gemfile and puppet 4 checks
In anticipation of puppet 4, start trying to deal with puppet 4 things
that can be helpfully predicted by puppet lint plugins. Fix arrow
alignment errors now caught by the bundle-installed puppet-lint.

This patch makes $users::virtual::localuser::sshkeys a required
parameter to fix the puppet-lint-empty_string-check error. If left as
empty string and the default is used, the ssh_authorized_key resource
will still create the authorized_key entry but with an empty key, which
is not useful.

Change-Id: Ica35f012f3af74159fc8b2ae5df60d11d4c15f33